Initial user model and account management flow
Sets default email host based on app name in PR builds

app.json now defines `ENV['PR_BUILD']`. If this is set, and heroku provides `HEROKU_APP_NAME`, construct a useful default host for emails.

We don't want to rely on this same construct in production as we'll eventually have a custom domain so just explicitly setting the domain in stage / prod is better. We can't reliably do that in PR builds as the host is dynamically generated when the build spins up.