Use new exact_value title matches to boost results
This boosts a lot on exact title matches, and a bit on any title match, and even less on contrbutor matches (non-exact).

The specific values are not well tuned yet, but are a good baseline starting point. These will likely move to configuration later, but for now we have magic numbers in code. Yay.