Update openapi to show optional security
This updates the openapi spec to show the optional security of "none". However, most documentation generators have no idea what to do with this even if it is officially part of the spec.

This also fixes the live swaggerui docs that load via github pages to use the json file and not fail looking for yaml we changed months ago as well as adding to additional generators so we can discuss them.

timdex.mit.edu still points to stoplight as our official documentation but as we've seen problems specifically around the authentication options we are exploring other options and moving these to our repo will allow others to see the docs live based on our current spec.