Refactor Retrieve to work with multiple promoted indexes
This has no user facing changes.

To date, we've only had one index "in production" at a time as we create or re-use the aleph_ indexes based on whether it is a full or partial ingest.

Now that we are adding records from an additional source, we'll always have more than one production index... one for each source.

This caused problems in how elasticsearch expects calls to retrieve a record work in that they require the index name to be passed, not the alias, in this situation.

This code avoids that problem by doing a search instead of a retrieved when asking for a single record. It's kinda dumb, but it seems fine.