Update dependencies and change default branch
Why these changes are being introduced:

Several dependencies have updates available, and we haven't yet
changed this repo's default branch to main.

Relevant ticket(s):

N/A

How this addresses that need:

This changes the default branch to main in GitHub Actions and
updates the following dependencies:
* puma: 5.0.2 to 5.1.0
* simplecov: 0.19.0 to 0.20.0
* rubocop: 0.92.0 to 1.4.2
* webmock: 3.9.1 to 3.10.0
* bootsnap: 1.4.8 to 1.5.1
* aws-sdk-elasticsearchservice: 1.43.0 to 1.56.0
* graphql: 1.11.5 to 1.11.6
* timecop: 0.9.1 to 0.9.2
* faraday_middleware-aws-sigv4: 0.3.0 to 0.4.0

Side effects of this change:

Just the residual effects of changing the main branch, which have
been addressed.