Update to Rails 6.0
Followed the process as defined:
https://edgeguides.rubyonrails.org/upgrading_ruby_on_rails.html

Includes `bin/rails app:update` and then review and adjustments to
changed files to match what we need.

Follow on work after we confirm this is stable should be done to move
the app to the new framework defaults by uncommenting the appropriate
lines in the initializer.