Provides end user documentation via GitHub pages
Generating on-the-fly using Swagger UI and the version of the api spec
in the main branch of the repository.

We will want to consider options to provide more control of which
version of the API spec we use in case we merge changes into main prior
to launching the changes into production but we currently need to get
this solution into production as our current solution was deleted by
the vendor with no notice.