Return remaining V2 fields
Why these changes are being introduced:

The data model includes fields that are not yet returned in V2
of the API.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/RDI-89

How this addresses that need:

This returns the additional V2 fields.

Side effects of this change:

* These fields have been added to the base set of fields, but this
may require modification once we determine the extended set.
* The following fields included in this commit are not required for
RDI: `contents`, `call_numbers`, `holdings`, `literary_form`,
`numbering`, `physical_description`, `publication_frequency.
* Cassettes have been regenerated.
* We are continuing to map `timdex_record_id` to `id`, based on
discussion in the PR. We may need to revisit this later on if it proves
confusing either to us or our users.