GraphQL backed by OpenSearch
Why are these changes being introduced:

* we are moving to GraphQL as our sole endpoint
* we are using OpenSearch as our backend

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/RDI-101

How does this address that need:

* Adds new v2 fields to GraphQL Records
* Deprecates fields by both documenting them as
  deprecated and which field to use instead, and also copies the data
  to the old field when possible
* Updates CI to run both OpenSearch and ElasticSearch versions of the
  tests

Document any side effects to this change:

note: filtering is mostly broken still at the Opensearch model level

note: Flipflip was used as it makes it slightly easier to flip states
in testing. I considered just using ENV but it was a bit clunky.

note: a few fields were unable to be deprecated cleanly. We could
consider renaming the new fields to allow for deprecation.

The tests need to be run twice until a better solution is identified or
we remove the elasticsearch version of the GraphQL (which will be in
the next few months). This is caused by our GraphQL implementation only
loading the config the first time it is used. The effect is that if an
elasticsearch test is run first, all the OpenSearch tests fail or vice
versa. The workaround was to name all of the tests as `graphqlv1` or
`graphqlv2` and using a filter to exclude running them. CI will run both
versions of the tests separately. Locally, developers should choose to
run whichever makes the most sense for their context. I wasn't able to
come up with a great way to default to one locally yet.