Configure Opensearch model to search single fields
Why these changes are being introduced:

The search model needs to be able to target any of the fields defined
as single-field searchable in v2.0 of the data model.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/RDI-102

How this addresses that need:

This enables matching on the following fields in the Opensearch model:

* citation
* contributors
* funding_information
* identifiers
* locations
* subjects

(Note that title was already made searchable as part of a previous
commit.)

Side effects of this change:

* These fields are not yet enabled in GraphQL, so you'll need to use
REST to confirm this behavior.
* While this allows us to search a single nested subfield, it does not
it does not allow for searching multiple nested subfields. For example,
searching contributors should search both 'value' and 'identifier', but
right now we can only search 'value'. RDI-116 will confirm expectations
on which subfields must be searchable.
* The way this targets individual fields seems to require an exact match.
This may be resolvable for nested fields by using `multi_match`, which
would also allow us to search multiple subfields. We plan to explore
this more as part of the graphql work.