Add targeted field search to GraphQL V2
Why these changes are being introduced:

The data model specifies that certain fields must be
single-field searchable.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/RDI-103

How this addresses that need:

This makes the following fields searchable in GraphQL:

* citation
* contributors
* funding_information
* identifiers
* location
* subjects
* title

This also renames the facet fields to avoid conflicts (see
side effects for more information on this).

Side effects of this change:

_This introduces a breaking change_. There is some overlap
between the fields above and facets. In order to accommodate both
searching and faceting, '_facet' has been added to all of the
facets in the QueryType class, and the corresponding change has
been made in the Opensearch model.

This means that anyone using the previous syntax for facets will
see their queries fail. We will want to do more analysis of how
the API is used before deploying this to prod. If we see usage
of faceting in prod, we might consider renaming the search fields
instead of the facets.