Return query match highlights and scores in GraphQL
Why these changes are being introduced:

Since we show minimal metadata in the brief results, it might be
useful for end users to see where in the metadata their search terms
matched.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/RDI-156

How this addresses that need:

This enables the OpenSearch feature to highlight query matches and
updates the GraphQL configuration to return that in the results.
Because we target `_source` as our GraphQL records, and the
highlights appear in `hits`, we inject the field into each record
in the QueryType class.

This approach also allows to return `score` in GraphQl results. Because
`score` will be useful to us in the near future, this commit configures
that as well.

Side effects of this change:

* We will likely want to reevaluate how we do this after removing
the deprecated V1 fields. I've added a comment to remind of this to
QueryType#inject_hits_fields_into_source, but we should also open a
JIRA ticket once we confirm that this approach is acceptable in the
interim.
* Some unrelated test assertions had to be updated, presumably
because more data sources had been loaded since the last time
cassettes were regenerated. If this becomes burdensome, we might
want to consider using staging OpenSearch and filtering secrets.

co-authored-by: Jeremy Prevost <jprevost@mit.edu>