Update stack to Ubuntu 22 and Heroku 22
Why these changes are being introduced:

We want to update our applications to use the Heroku 22
stack, which also requires us to use Ubuntu 22.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ENGX-180
https://mitlibraries.atlassian.net/browse/ENGX-182

How this addresses that need:

This updates the CI config to point to the shared workflow
and adds the Heroku stack version to app.json.

Side effects of this change:

* Updates webmock to the latest version, as the version
currently installed has been removed from rubygems.
* We cannot use the shared CI workflow with TIMDEX yet, since we
need to run V1 and V2 tests separately. Once we retire V1, we
should update this repo to use the shared workflow to make it
consistent with our other Rails apps. I opened TIMX-84 to
track this.