Upgrade to Rails 7
Why these changes are being introduced:

We want to stay up to date with the latest Rails.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ENGX-155

How this addresses that need:

This upgrades to Rails 7 and loads the new framework defaults.
Prior to this, I ran `rails app:update`, checked the diffs for all
files, and confirmed that the new framework defaults would not
cause problems. In this case, it seemed like there were no Rails
files that needed to be updated, so I kept the existing files.

Side effects of this change:

* The way we handle domain redirects
(`ApplicationController#ensure_domain`) now throws an unsafe
redirect error. I've added the `allow_other_host` param to that
method to bypass this. In this case the other host would come
from env, so the risk seems low, but I'd encourage the reviewer
to consider any edge cases I may have missed.
* The only new framework default that seemed like it could cause
problems for us was the key generator digest changing to SHA256:
https://guides.rubyonrails.org/v7.0/upgrading_ruby_on_rails.html#key-generator-digest-class-changing-to-use-sha256.
However, However, I don't think we are using encrypted cookies
in this app, so I don't anticipate problems there.