Adds skeletal jekyll based user facing docs
Why are these changes being introduced:

* We'll be creating documentation using this process and this will
  allow us to test the infrastructure with a single page before moving
  on

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/TIMX-148

How does this address that need:

* Adds basic configuration for jekyll using the remote theme
  "just-the-docs" and a single "index_next" page.

Document any side effects to this change:

This adds 3 new development dependencies to the timdex-api. This is a
convenience to make it easier to make sure all developers can easily
run the jekyll site and are using the same version of the libraries. It
is also, however, not-ideal to add these libraries to the timdex-api
application. Since they are dev-only, it feels acceptable even if it is
not ideal.