Add descriptions to GraphQL V2 fields
Why these changes are being introduced:

Many of the GraphQL fields are not documented.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/TIMX-116

How this addresses that need:

This adds descriptions to all V2 fields and search arguments, using
the TIMDEX Data Model v2.0 document as a guide.

Side effects of this change:

* Descriptions have been modified from the data model where
applicable, or newly written where unavailable (specifically,
subfields and aggregations).
* One field, 'numbering', does not have a description in the data
model. I've set this aside until we have a better understanding
of what it does.