Holdings.location field should be optional
Why these changes are being introduced:

Many records may not have a holdings location. However, this field
is currently non-nullable, so TIMDEX throws an error when it is
missing.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/TIMX-171

How this addresses that need:

This makes holdings.location a nullable field.

Side effects of this change:

None.