Fix default index for V2 retrieve queries
Why these changes are being introduced:

The Retrieve model uses the default elasticsearch index name for
V1 and V2 queries, which means that V2 retrieve queries use the
wrong index name.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/TIMX-185

How this addresses that need:

This updates Retrieve#default_index to use the default opensearch
index name for V2 queries.

Side effects of this change:

The regression tests for this pass, but somewhat disingenously as
we currently use the same index names in .env.test for elasticsearch
and opensearch. We should probably regenerate cassettes with the
latest prod data and start using the `all-current` index for V2
queries.