Add initial how-to guides
Why these changes are being introduced:

Version 2 of the API introduces significant changes that should
be documented publicly for our end users.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/TIMX-192

How this addresses that need:

This adds an initial set of how-to guides, which provide specific
example queries of certain features.

Side effects of this change:

* There are some WAVE errors, primarily related to low contrast.
We might want to use the light color scheme when we launch.
* Since this is not considered an external link like the
SpectaQL-generated reference docs, it seems to always appear in
the nav bar above reference, no matter how the nav order is set.
This is probably fine, but if I were developing an app with TIMDEX,
the reference docs are probably what I would look up most frequently
(although maybe not first).