Link to V2 dev docs from splash page
Why these changes are being introduced:

There is no clear path to the V2 docs from our splash page.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/TIMX-209

How this addresses that need:

This adds a link to the V2 dev docs on the splash page.

Side effects of this change:

* The link in the deprecated docs paragraph now goes to the V2 docs,
so I updated that to go to the deprecated docs page.
* The db schema has been updated to include Rails 7 changes. (This
PR does not require db migrations, but I had to run them because I
am using a fresh clone of the repo.)