Fix typo on splash page
Why these changes are being introduced:

The last update to the docs added a typo that I missed at the time,
but will make me flip a table every time I see it.

Relevant ticket(s):

N/A.

How this addresses that need:

This fixes the typo.

Side effects of this change:

None.