Update graphql-ruby and specify LegacyTrace class
Why these changes are being introduced:

The tracer API was overhauled in v2.0.18 of GraphQL for
performance reasons. Custom tracers are now defined as modules
instead of classes, and hooks are mixed in using `Schema.trace_with`.
The old way of defining custom tracers is still allowed by
adding `trace_class(GraphQL::Tracing::LegacyTrace)` to the schema.

See https://github.com/rmosolgo/graphql-ruby/pull/4344 for more
info on this change.

Relevant ticket(s):

N/A.

How this addresses that need:

This updates graphql-ruby to the latest version and specifies the
legacy tracer API in the schema.

Side effects of this change:

We should update our tracer to use the new API. I opened a ticket
to that effect: https://mitlibraries.atlassian.net/browse/ENGX-221