Update theme gem
Why these changes are being introduced:

The new theme gem version includes some footer updates requested by UXWS.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/UXWS-1507

How this addresses that need:

This updates the theme gem.

Side effects of this change:

The previous version missed the `versions.rb` change, so the lockfile will show an increment from `1.0.0` to `1.0.2`, when in reality it is from `1.0.1` to `1.0.2`.