Update opensearch
Why these changes are being introduced:

We had pinned opensearch to v2.1 in the gemfile due to an
upstream bug in more recent versions. That bug has been fixed,
so we can now update the gem.

Relevant ticket(s):

N/A.

How this addresses that need:

This removes the version specification for opensearch and
updates the gem.

Side effects of this change:

* The test suite now generates `test.sqlite3-shm` and
`test.sqlite3-war` files. I've updated the gitignore to ignore
these (and any suffixed `test.sqlite3` files).
* The psych gem has been updated, unrelated to the opensearch
update.