Adds places aggregations
Why are these changes being introduced:

* geo ui has reqeusted the ability to filter on additional fields

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/GDT-168
* https://mitlibraries.atlassian.net/browse/GDT-169

How does this address that need:

* Adds aggregation in OpenSearch for places by utilizing `Subjects` with
  a `kind` of `Dublin Core; Spatial`

Not included

* Aggregations for `Access Type` and `Data Type` are not yet included
  as we are still working out exactly how that data should be stored as
  well as `years` which which will also be handled separately

Document any side effects to this change:

The default index we used in tests was changed from the old value to
`all-current` which should make it easier to generate new tests and
cassettes in the future as the old value (`timdex-prod`) was no longer
used anywhere but in our tests.

Changing aggregations required rebuilding most of our cassettes. While I
was rebuilding most of them, I opted to rebuild them all. This includes
sensitive data scrubbing automatically when using AWS opensearch
instances (srubs credentials and URI of the instance). The README has
instructions on how to generate cassettes successfully.

I generated cassettes from our current dev1 data, which means a lot of
explicitly checked for values changed in the tests. While updating tests
for value changes, I also took this opportunity to remove most of the
remnants from the v2 naming we had used during the transition from v1.
I have not yet renamed the test file, as I was worried that would be
more complicated of a review. A follow on change to just rename that
file will be forthcoming.