Add publishers field to record type
Why these changes are being introduced:

We recently [made the decision in transmogrifier](https://github.com/MITLibraries/transmogrifier/pull/132)
To map publication information to a new nested field, named
`publishers` to avoid a naming conflict with the existing
`publicationInformation` field.

Relevant ticket(s):

* [GDT-270](https://mitlibraries.atlassian.net/browse/GDT-270)
* [GDT-218](https://mitlibraries.atlassian.net/browse/GDT-218)
* [GDT-229](https://mitlibraries.atlassian.net/browse/GDT-229)
* [GDT-271](https://mitlibraries.atlassian.net/browse/GDT-271)

How this addresses that need:

This adds the new `publishers` field and deprecates
`publicationInformation`.

Side effects of this change:

A follow-on ticket is needed to add this field to the UI. (This
is ticketed as GDT-271.)